Alterado de JDBC para JPA com Mysql 5.1
