+++ README.md
@@ -3,3 +3,7 @@
 Certificado Alura: SPRING MVC: É HORA DE CRIAR UMA WEBAPP COM SPRING MVC4
 
 https://cursos.alura.com.br/certificate/05d86f52-be97-4b4d-9f25-036f95e7be53
+
+# Melhorias
+
+Alterado de JDBC para JPA e alterado o banco para MYSQL
