+++ WebContent/WEB-INF/spring-context.xml
@@ -3,12 +3,15 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:context="http://www.springframework.org/schema/context"
 	xmlns:mvc="http://www.springframework.org/schema/mvc"
+	xmlns:tx="http://www.springframework.org/schema/tx"
 	xsi:schemaLocation="http://www.springframework.org/schema/mvc 
 						http://www.springframework.org/schema/mvc/spring-mvc-3.0.xsd
 						http://www.springframework.org/schema/beans 
 						http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
 						http://www.springframework.org/schema/context 
-						http://www.springframework.org/schema/context/spring-context-3.0.xsd">
+						http://www.springframework.org/schema/context/spring-context-3.0.xsd
+						http://www.springframework.org/schema/tx 
+						http://www.springframework.org/schema/tx/spring-tx-3.0.xsd ">
 
 	<context:component-scan
 		base-package="br.com.caelum.contas" />
@@ -17,7 +20,10 @@
 	<mvc:interceptors>
 		<bean class="br.com.caelum.contas.interceptor.AutorizadorInterceptor"></bean>
 	</mvc:interceptors>
-
+<!-- 	<tx:annotation-driven transaction-manager="transactionManager"/>  -->
+	<!-- Serve para ativar a anotação @Transaction  -->
+	<tx:annotation-driven/>
+    
 	<bean
 		class="org.springframework.web.servlet.view.InternalResourceViewResolver">
 		<property name="prefix" value="/WEB-INF/views/" />
@@ -33,11 +39,34 @@
 		</property>
 	</bean>
 	
-	<bean id="mysqlDataSource" class="org.apache.commons.dbcp.BasicDataSource">
+ <!--  	utilizado no JDBC e HSQL -->
+<!--	<bean id="hsqlDataSource" class="org.apache.commons.dbcp.BasicDataSource">
 		<property name="driverClassName" value="org.hsqldb.jdbcDriver" />
 		<property name="url" value="jdbc:hsqldb:file:contas.db" />
 		<property name="username" value="sa" />
 		<property name="password" value="" />
-	</bean>	
+	</bean>	 -->
+	
+	<!--  	utilizado no JPA e MYSQL -->
+	<bean id="mysqlDataSource" class="org.apache.commons.dbcp.BasicDataSource">
+	    <property name="driverClassName" value="com.mysql.jdbc.Driver"/>
+	    <property name="url" value="jdbc:mysql://localhost/contasSpring?createDatabaseIfNotExist=true"/>
+	    <property name="username" value="root"/>
+	    <property name="password" value="root"/>
+	</bean>
+	
+	<!-- Utilizado no JPA  -->
+	<bean id="entityManagerFactory" 
+        class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
+	    <property name="dataSource" ref="mysqlDataSource" />
+	    <property name="jpaVendorAdapter">
+	        <bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter"/>
+	    </property>
+	</bean>
 
+	<bean id="transactionManager" 
+	            class="org.springframework.orm.jpa.JpaTransactionManager">
+	    <property name="entityManagerFactory" ref="entityManagerFactory"/>
+	</bean>
+	
 </beans>
