+++ src/META-INF/persistence.xml
@@ -1,0 +1,26 @@
+<persistence xmlns="http://java.sun.com/xml/ns/persistence"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://java.sun.com/xml/ns/persistence
+     http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
+    version="2.0">
+
+ <persistence-unit name="contas">
+
+   <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+
+   <class>br.com.caelum.contas.modelo.Conta</class>
+   <class>br.com.caelum.contas.modelo.Usuario</class>
+
+   <properties>
+
+    <!-- SEM as propriedades URL, login, senha e driver, pois estao no arquivo spring-context.xhtml -->
+
+    <property name="hibernate.dialect" 
+            value="org.hibernate.dialect.MySQL5InnoDBDialect" />
+    <property name="hibernate.show_sql" value="true" />
+    <property name="hibernate.format_sql" value="true" />
+    <property name="hibernate.hbm2ddl.auto" value="update" />        
+   </properties>
+ </persistence-unit>
+
+</persistence>
