+++ src/br/com/caelum/contas/controller/InfraController.java
@@ -4,15 +4,24 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.com.caelum.contas.ConnectionFactory;
+import br.com.caelum.contas.dao.UsuarioDAO;
+import br.com.caelum.contas.modelo.Usuario;
 
 @Controller
 public class InfraController {
-
-	@RequestMapping("/tabelas")
+	
+	@Autowired
+	@Qualifier("jpaUsuarioDAO")
+	private UsuarioDAO dao;
+	
+	@RequestMapping("/tabelasJdbc")
 	public String criaBanco() throws SQLException {
 		Connection c = new ConnectionFactory().getConnection();
 		PreparedStatement st1 = c.prepareStatement("drop table contas if exists");
@@ -35,4 +44,20 @@ public class InfraController {
 		return "infra-ok";
 
 	}
+	
+	@Transactional
+	@RequestMapping("/criaUsuario")
+	public String criaUsuario() {
+		Usuario usuario = new Usuario();
+		if (!dao.existeUsuario(usuario)) {
+			usuario.setLogin("caelum");
+			usuario.setSenha("online");
+			dao.insere(usuario);
+		}
+		
+		return "infra-ok";
+		
+	}
+	
+	
 }
