+++ src/br/com/caelum/contas/controller/UsuarioController.java
@@ -2,6 +2,8 @@ package br.com.caelum.contas.controller;
 
 import javax.servlet.http.HttpSession;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -12,6 +14,10 @@ import br.com.caelum.contas.modelo.Usuario;
 @Controller
 public class UsuarioController {
 
+	@Autowired
+	@Qualifier("jpaUsuarioDAO")
+	private UsuarioDAO dao;
+	
 	@RequestMapping("/loginForm")
 	public String login() {
 		return "usuario/login";
@@ -20,7 +26,6 @@ public class UsuarioController {
 	@RequestMapping("/efetuaLogin")
 	public String efetuaLogin(Usuario usuario, HttpSession session, RedirectAttributes redirectAttributes) {
 
-		UsuarioDAO dao = new UsuarioDAO();
 		if (dao.existeUsuario(usuario)) {
 			session.setAttribute("usuarioLogado", usuario);
 			return "menu";
