+++ src/br/com/caelum/contas/dao/UsuarioDAO.java
@@ -1,60 +1,11 @@
 package br.com.caelum.contas.dao;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import br.com.caelum.contas.ConnectionFactory;
 import br.com.caelum.contas.modelo.Usuario;
 
-public class UsuarioDAO {
-	private Connection connection;
+public interface UsuarioDAO {
 
-	public UsuarioDAO() {
-		try {
-			connection = new ConnectionFactory().getConnection();
-		} catch (SQLException e) {
-			throw new RuntimeException(e);
-		}
-	}
+	boolean existeUsuario(Usuario usuario);
 
-	public boolean existeUsuario(Usuario usuario) {
-		
-		if(usuario == null) {
-			throw new IllegalArgumentException("Usuário nao deve ser nulo");
-		}
-		
-		try {
-			PreparedStatement stmt = this.connection.prepareStatement("select * from usuarios where login = ? and senha = ?");
-			stmt.setString(1, usuario.getLogin());
-			stmt.setString(2, usuario.getSenha());
-			ResultSet rs = stmt.executeQuery();
+	void insere(Usuario usuario);
 
-			boolean encontrado = rs.next();
-			rs.close();
-			stmt.close();
-			
-			return encontrado;
-		} catch (SQLException e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	public void insere(Usuario usuario) {
-		if(usuario == null) {
-			throw new IllegalArgumentException("Usuário nao deve ser nulo");
-		}
-		
-		try {
-			PreparedStatement stmt = this.connection.prepareStatement("insert into usuarios (login,senha) values (?,?)");
-			stmt.setString(1, usuario.getLogin());
-			stmt.setString(2, usuario.getSenha());
-			stmt.execute();
-
-		} catch (SQLException e) {
-			throw new RuntimeException(e);
-		}		
-	}
 }
-
